Add tests for new survey view
