+++ Pipfile
@@ -21,6 +21,7 @@ jsmin = "==2.2.2"
 gunicorn = "==19.7.1"
 scss = "*"
 sass = "*"
+requests-mock = "*"
 
 
 [dev-packages]
