+++ tests/test_data/survey/survey_list.json
@@ -1,0 +1,14 @@
+[
+  {
+    "id": "cb0711c3-0ac8-41d3-ae0e-567e5ea1ef87",
+    "shortName": "BRES",
+    "longName": "Business Register and Employment Survey",
+    "surveyRef": "221"
+  },
+  {
+    "id": "cb0711c3-0ac8-41d3-ae0e-567e5ea1ef88",
+    "shortName": "BRUS",
+    "longName": "Business Register and Umployment Survey",
+    "surveyRef": "222"
+  }
+]
