+++ tests/views/test_info.py
